feat(/migrations): temp migrations for transfer items

+++ database/migrations/2024_03_12_053205_add_company_id_to_users_table.php
@@ -13,8 +13,11 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedBigInteger('company_id')->nullable()->after('id');
-            $table->foreign('company_id')->references('id')->on('companies')->onDelete('set null');
+            if (!Schema::hasColumn('users', 'company_id')) {
+
+                $table->unsignedBigInteger('company_id')->nullable()->after('id');
+                $table->foreign('company_id')->references('id')->on('companies')->onDelete('set null');
+            }
         });
     }
