+++ database/migrations/2024_03_14_072603_add_company_id_to_items_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('items', function (Blueprint $table) {
-            $table->foreignId('company_id')->after('id')->constrained()->onDelete('cascade');
+             if(!Schema::hasColumn('items', 'company_id')) {
+                $table->foreignId('company_id')->after('id')->constrained()->onDelete('cascade');
+             }
         });
     }
